α-replacement outsourced
to a separate function (called 'alpha_replacement'). Then an α-replacement is performed at the end of each β-reduction in order to prevent possible errors in future β-reductions. (For example, if beta_replacer_ST gets substituted into more than one location and it contains at least one λ, then the variable bound by that λ would show up in separate locations with non-overlapping scopes, and α-replacement would be applicable to at least one of the instances of that bound variable.)