adding whole variable-name safeguards for β
The α-replacement function already has a couple of safeguards (at L495, L498) to make sure that when a candidate replacee is detected, what is detected is actually a whole variable-name and not just part of it. Now those safeguards are being transferred (copied) to the β-reduction function, to L333 & L334, to vet the candidate β-replacees there.